+++ onnx/src/main/kotlin/org/jetbrains/kotlinx/dl/api/inference/onnx/posedetection/MultiPoseDetectionModel.kt
@@ -26,7 +26,7 @@ private const val INPUT_SIZE = 256
 /**
  * MultiPoseDetectionModel is an ultra-fast and accurate model that detects 6 persons with 17 keypoints and 18 basic edges of a body for each of them.
  *
- * It internally uses [ONNXModels.PoseEstimation.MoveNetMultiPoseLighting] under the hood to make predictions.
+ * It internally uses [ONNXModels.PoseDetection.MoveNetMultiPoseLighting] under the hood to make predictions.
  */
 public class MultiPoseDetectionModel : OnnxInferenceModel() {
     public fun detectPoses(inputData: FloatArray, confidence: Float = 0.005f): MultiPoseDetectionResult {
@@ -84,7 +84,7 @@ public class MultiPoseDetectionModel : OnnxInferenceModel() {
 
         val (data, shape) = preprocessing(imageFile)
 
-        val preprocessedData = ONNXModels.PoseEstimation.MoveNetSinglePoseLighting.preprocessInput(
+        val preprocessedData = ONNXModels.PoseDetection.MoveNetSinglePoseLighting.preprocessInput(
             data,
             longArrayOf(shape.width!!, shape.height!!, shape.channels!!)
         )